Fixed thorium duping via fluid extractor
The way nuclearcraft isotopes work is a natural ingot (144mb) is split into a clump (144mb) and small clump (16mb), both oredicted as a ingot and nugget. Gregtech sees these oredicts, and decides to add recipe in the fluid extractor for thorium, even though these aren't supposed to be recombined into natural thorium. The script removes these recipes, and adds a new one in the mixer to recombine the correct isotopes into thorium dust.