Shift quest IDs
Programmatically push all new quests to the end of the database to ensure that there will be no ID collisions.
This makes the new layout safe to apply to older worlds.

Get rid of several completely obsolete quests and hides the corresponding tab.

Fill gap IDs to prevent BetterQuesting from condesing the database.