Switch to Dimensional Edibles: Omnifactory Edition (#725)
Our fork of Dimensional Edibles fixes spawn-below-map and spawn-in-lava issues, as well as improvements to item tooltips and overlays. If updating from an older version, use the /de invalidate command to reset a bad spawn position for a target player and dimension. This command requires operator permissions in a dedicated server environment.

Co-authored-by: ALongStringOfNumbers <31759736+ALongStringOfNumbers@users.noreply.github.com>