Update Electronics.zs
I'm suggesting a change to the primitive Diode assembler recipe.

This is primarily done to avoid a conflict that arises when trying to craft SMD Diodes and SMD Transistors in the same machine.

Please note that the specific recipe is not finalized and probably requires some discussion if everyone agrees that changing the recipe is acceptable.

My thought behind the changes were:

Fine Annealed Copper Wire -> Fine Tin Alloy Wire - Annealed Copper happens a bit later in the chain than the time when you are manually crafting Primitive Processors (the only components that primitive Diodes are used for).  It seems more reasonable to me that you would replace the Tin Wire in the manual craft with a more advanced version of Tin, hence Tin Alloy.

288mb Polyethylene -> Removed - There are four primitive components, two of which have Rubber Sheets in the Crafting recipe (Transistor and Capacitor), and two of which don't (Diode and Resistor).  The Polyethylene in the recipe seems to replace the rubber and so in my opinion shouldn't be used to craft the Diode.  I would also suggest changing the Polyethylene sheet to liquid Polyethylene and adding a Silicon Plate to the Capacitor recipe, but that's a different issue.

MV Power -> HV Power - Crafting these in an Assembler should not be happening during progression.  Since better tier two circuits are crafted using transistors which are earlier in the progression chain, it seems that further incentivizing using those should help push the player in to quick circuit advancement.  The only real use case for the Diode Assembler recipe is for auto crafting of Primitive Processors in a Bounty Board setup, which at earliest happens at the very end of MV or during HV.  Adjusting the power requirements to match that seemed to make sense.