Add Multiblock Simulation Chamber: Simulation Supercomputer (#720)
Adds the Simulation Supercomputer, a multiblock structure that can overclock to iterate monster data models extremely fast for pristine and world matter. It will also train up models not yet at Self Aware, but overclocking DOES NOT affect the pristine chances (contrary to the tooltip in JEI).

Also adds a quest for the structure in the Endgame tab, depending on the Creative Tank quest. This structure is intended to be an early Creative Tier machine, requiring a heart but otherwise very easy to make with the help of the Creative Tank.

Co-authored-by: ALongStringOfNumbers <31759736+ALongStringOfNumbers@users.noreply.github.com>
Co-authored-by: Exaxxion <Exaxxion@users.noreply.github.com>

[skip ci]