Update Applied Energistics 2: Trousers Edition from v46r to v47
Note that this version of PAE2 reverts all changes Proto made to the AE2 energy net and the item counts caching, due to some players experiencing bugs. If you weren't having problems with v46, you should be able to safely manually switch back to one of those versions. We are updating to use v47 in the interests of stability.