Convert almost all fluid bucket item retrieval quests to fluid quests
Replaces all of the quests which exclusively request fluid in buckets to
instead simply request the fluid, allowing any suitable container to
qualify.

Quests with a mixture of items and fluid buckets are unchanged, to help
avoid confusion since mixed task types are on separate task pages within
the quest book, and it's not particularly obvious.

Only downside is that these quests might not sync automatically if a new
player joins the quest group after this change. Just pick up the fluids,
and it will complete normally.