Fix Canola Oil-based Glycerol recipes
For whatever reason the base recipe for making Glycerol using Canola
Oil had its quantities of each ingredient reversed relative to the
Glycerol recipes using Seed and Fish Oil.

This swaps the values to bring it back in line with the other recipes,
while still rewarding the player for refining the Canola Oil into
Refined, Crystallized, and Empowered forms.

[skip ci]