Rework Iridium quest into "Osmium and Iridium"
Since Iridium is really easy to obtain the quest didn't make much sense
as it was. Reworked the quest into one for Osmium, Iridium, Osmiridium.

Removed several dependencies on Tier Four Micro Miner in favor of this
new quest instead.

Added dependency on this quest to Conflux Catalyst and Gem Sensor.

Added a redundant dependency on it by Ultimate Extended Crafting Table
so the PAuto/PExC quests aren't stranded on the page.